requirements.txt
pillow
requests
numpy
panel
xarray
holoviews
bokeh
